refactor(queries): move process, census and draft hooks into the queries layer
