+++ src/elements/dashboard/memberbase/members.tsx
@@ -1,60 +1,13 @@
 import { RoutedPaginationProvider } from '@vocdoni/react-components/pagination'
-import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useOutletContext } from 'react-router-dom'
 import { ListStateAlert } from '~components/Feedback/ListStateAlert'
 import { MemberbaseTabsContext } from '~components/Memberbase'
 import MembersTable from '~components/Memberbase/Members'
 import { TableProvider } from '~components/Memberbase/TableProvider'
+import { useMemberColumns } from '~components/Memberbase/useMemberColumns'
 import { Routes } from '~routes'
 import { usePaginatedMembers } from '~src/queries/members'
-
-export const useMemberColumns = () => {
-  const { t } = useTranslation()
-
-  return useMemo(
-    () => [
-      {
-        label: t('members.fields.firstname', { defaultValue: 'First Name' }),
-        id: 'name',
-      },
-      {
-        label: t('members.fields.surname', { defaultValue: 'Last Name' }),
-        id: 'surname',
-      },
-      {
-        label: t('members.fields.email', { defaultValue: 'Email' }),
-        is2fa: true,
-        id: 'email',
-      },
-      {
-        label: t('members.fields.phone', { defaultValue: 'Phone' }),
-        is2fa: true,
-        id: 'phone',
-      },
-      {
-        label: t('members.fields.member_number', { defaultValue: 'Member Number' }),
-        id: 'memberNumber',
-      },
-      {
-        label: t('members.fields.national_id', { defaultValue: 'National ID' }),
-        id: 'nationalId',
-        visible: false,
-      },
-      {
-        label: t('members.fields.birth_date', { defaultValue: 'Birth Date' }),
-        id: 'birthDate',
-        visible: false,
-      },
-      {
-        label: t('members.fields.weight', { defaultValue: 'Voting power (Weight)' }),
-        id: 'weight',
-        visible: true,
-      },
-    ],
-    [t]
-  )
-}
 
 const Members = () => {
   const { t } = useTranslation()
